+++ reach/rule_match.go
@@ -1,0 +1,6 @@
+package reach
+
+type RuleMatch interface {
+	Explain(observedDescriptor string) Explanation
+	GetRule() *SecurityGroupRule
+}
