+++ reach/security_group_rule.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/logrusorgru/aurora"
 	"github.com/luhring/reach/network"
 	"net"
 )
@@ -78,51 +77,4 @@ func (rule *SecurityGroupRule) matchWithInterface(targetInterface *NetworkInterf
 	}
 
 	return nil
-}
-
-type SGRefRuleMatch struct {
-	Rule  *SecurityGroupRule
-	SGRef *SecurityGroupReference
-}
-
-func (m *SGRefRuleMatch) Explain(observedDescriptor string) Explanation {
-	var explanation Explanation
-
-	explanation.AddLineFormat("security group (%v)", m.SGRef.Name)
-
-	return explanation
-}
-
-type IPRuleMatch struct {
-	Rule             *SecurityGroupRule
-	MatchedIPRange   *net.IPNet
-	TargetIP         net.IP
-	IsTargetIPPublic bool
-}
-
-func (m *IPRuleMatch) Explain(observedDescriptor string) Explanation {
-	var explanation Explanation
-
-	var publicOrPrivate string
-	if m.IsTargetIPPublic {
-		publicOrPrivate = "public"
-	} else {
-		publicOrPrivate = "private"
-	}
-
-	explanation.AddLineFormatWithEffect(aurora.Green, "- rule: allow %v", aurora.Bold(m.Rule.TrafficAllowance.Describe()))
-	explanation.AddLineFormatWithIndents(
-		1,
-		"(This rule handles an IP address range '%v' that includes the %s network interface's %s IP address '%v'.)",
-		m.MatchedIPRange.String(),
-		observedDescriptor,
-		publicOrPrivate,
-		m.TargetIP.String(),
-	)
-
-	return explanation
-}
-
-type RuleMatch interface {
-	Explain(observedDescriptor string) Explanation
 }
